# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: API de gestión de incidencias
  version: 0.1.0
  description: >-
    Contrato inicial de la plataforma. Las operaciones se implementarán por
    iteración; publicar este archivo no implica que los endpoints estén activos.
...
servers:
  - url: http://localhost:3000/api/v1
    description: Desarrollo local (desde Iteración 1)
...
paths:
...
  /incidents/{id}/evidence:
    post:
      tags: [Incidents]
      summary: Adjuntar fotografía al reporte
      operationId: uploadIncidentEvidence
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        '201':
          description: Fotografía almacenada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceMetadata'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /incidents/{id}/evidence/{evidenceId}:
    get:
      tags: [Incidents]
      summary: Leer fotografía privada y auditar acceso
      operationId: readIncidentEvidence
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IncidentId'
        - { name: evidenceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Bytes de imagen autorizada, sin URL pública estable
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
            image/webp: { schema: { type: string, format: binary } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /catalog/categories:
    get:
      tags: [Catalog]
      operationId: listIncidentCategories
      responses:
        '200':
          description: Categorías activas con prioridad inicial
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IncidentCategory' } }
    post:
      tags: [Catalog]
      operationId: createIncidentCategory
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IncidentCategoryInput' }
      responses:
        '201':
          description: Categoría creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncidentCategory' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /catalog/categories/{id}:
    patch:
      tags: [Catalog]
      operationId: updateIncidentCategory
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IncidentCategoryInput' }
      responses:
        '200':
          description: Categoría actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncidentCategory' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /catalog/categories/{id}/subcategories:
    get:
      tags: [Catalog]
      operationId: listIncidentSubcategories
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      responses:
        '200':
          description: Subcategorías activas
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IncidentSubcategory' } }
    post:
      tags: [Catalog]
      operationId: createIncidentSubcategory
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IncidentSubcategoryInput' }
      responses:
        '201':
          description: Subcategoría creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncidentSubcategory' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /catalog/subcategories/{id}:
    patch:
      tags: [Catalog]
      operationId: updateIncidentSubcategory
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IncidentSubcategoryInput' }
      responses:
        '200':
          description: Subcategoría actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncidentSubcategory' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string }
    IncidentId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Autenticación requerida o inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Acceso fuera del rol o ámbito autorizado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Conflicto de estado o duplicidad
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    PayloadTooLarge:
      description: Fotografía supera el límite de tamaño
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    UnprocessableEntity:
      description: Contraseña fuera de la política de seguridad
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServiceUnavailable:
      description: Servicio de autenticación o recuperación no configurado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    IncidentCategoryInput:
      type: object
      required: [code, name, family]
      properties:
        code: { type: string, maxLength: 60 }
        name: { type: string, maxLength: 120 }
        family: { type: string, enum: [emergency, security] }
        defaultPriority: { type: string, enum: [low, normal, high, critical] }
        isSensitive: { type: boolean }
        active: { type: boolean }
    IncidentCategory:
      allOf:
        - { $ref: '#/components/schemas/IncidentCategoryInput' }
        - type: object
          required: [id, defaultPriority, active, isDemo]
          properties:
            id: { type: integer }
            isDemo: { type: boolean }
    IncidentSubcategoryInput:
      type: object
      required: [code, name]
      properties:
        code: { type: string, maxLength: 60 }
        name: { type: string, maxLength: 120 }
        isSensitive: { type: boolean }
        active: { type: boolean }
    IncidentSubcategory:
      allOf:
        - { $ref: '#/components/schemas/IncidentSubcategoryInput' }
        - type: object
          required: [id, categoryId, active, isDemo]
          properties:
            id: { type: integer }
            categoryId: { type: integer }
            isDemo: { type: boolean }
...
    ApiError:
      type: object
      required: [code, message, correlationId]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        correlationId: { type: string }
...
    EvidenceMetadata:
      type: object
      required: [id, mediaType, uploadedAt]
      properties:
        id: { type: string }
        mediaType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        uploadedAt: { type: string, format: date-time }
...